box/journal: use plain int for return value
We're returning int64_t with values 0 or -1 by now,
there is no need for such big return type, plain
integer is enough.

Acked-by: Nikita Pettik <korablev@tarantool.org>

+++ src/box/box.cc
@@ -310,7 +310,7 @@ struct recovery_journal {
  * exact same signature during local recovery to properly mark
  * min/max LSN of created LSM levels.
  */
-static int64_t
+static int
 recovery_journal_write(struct journal *base,
 		       struct journal_entry *entry)
 {
